requirements.txt
azure-functions==1.20.0
fastapi==0.111.0
uvicorn==0.30.1
